Apply suggestions from ARO-RP code review
